backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dominik Mack <dominik.mack@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
chalice = "^1.29.0"
boto3 = "^1.28.73"
pydantic = "^2.4.2"
pytest-postgresql = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
